src/spicesimulation.py: fix https://github.com/mawatfa/ebana/issues/4
The __init__ method is fixed to something which looks more logical but
the initial idea of the author remains unclear.

Now the attribute is initialized to ngspice when it is not set.

When the attribute is set to xyce specifically, the DEFAULT_SIMULATOR
attribute in PySpice is set to a specific value